Split Health tab into lazily-loaded sub-sections; harden SQL scans for large tenants
The in-app Health page loaded one monolithic HealthDashboard (SQL + schema + config + Graph + 7 App Insights queries) under a single lock. It's now split into independently-cached per-section endpoints (summary/data/liveness/exceptions/components/config); the SPA fetches only the sub-section the user opens via lazily-mounted tabs, with the Overview showing the rolled-up traffic-light.

Performance: audit_events.time_stamp and hits.hit_timestamp are unindexed, so the data section now uses a short per-query CommandTimeout, scans the two big fact tables in parallel folding 24h/7d counts + freshness into a single pass, and degrades to a per-metric note instead of hanging the request. The Overview roll-up skips the heavy SQL entirely (SELECT 1 reachability probe).
