+++ src/AnalyticsEngine/Web/Scripts/admin-app/index.html
@@ -28,6 +28,7 @@
     window.o365AnalyticsSystemStatusAPI = baseUrl + "api/SystemStatus";
     window.o365AnalyticsInstallLogAPI = baseUrl + "api/InstallLog";
     window.o365AnalyticsProfilingStatusAPI = baseUrl + "api/ProfilingStatus";
+    window.o365AnalyticsHealthAPI = baseUrl + "api/Health";
   </script>
 </head>
